Probes Configuration Updater & Comparer
Probes Configuration Updater & Comparer implementation

+++ tracer/src/Datadog.Trace/Debugger/Configurations/Models/OpsConfiguration.cs
@@ -7,5 +7,5 @@ namespace Datadog.Trace.Debugger.Configurations.Models;
 
 internal class OpsConfiguration
 {
-    public long PollInterval { get; set; }
+    public int PollInterval { get; set; }
 }
